Removed statsThread, moved VRS stuff into VrsService
