+++ PlaneAlerter/Services/CheckerService.cs
@@ -2,25 +2,20 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Globalization;
-using System.IO;
 using System.Linq;
 using System.Media;
-using System.Net;
-using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Windows.Forms;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using PlaneAlerter.Enums;
 using PlaneAlerter.Models;
+using Match = PlaneAlerter.Models.Match;
 
 namespace PlaneAlerter.Services {
 	internal interface ICheckerService
 	{	
 		void Start();
 		void Stop();
-
-		Dictionary<string, string>? GetReceivers();
 	}
 
 	/// <summary>
@@ -30,25 +25,16 @@ namespace PlaneAlerter.Services {
 	{
 		private readonly ISettingsManagerService _settingsManagerService;
 		private readonly IConditionManagerService _conditionManagerService;
+		private readonly IVrsService _vrsService;
 		private readonly IEmailService _emailService;
 		private readonly ITwitterService _twitterService;
 		private readonly IUrlBuilderService _urlBuilderService;
 		private readonly IStatsService _statsService;
 
 		/// <summary>
-		/// Client for sending aircraftlist.json requests
-		/// </summary>
-		private HttpWebRequest _request;
-
-		/// <summary>
 		/// Time of next check
 		/// </summary>
 		private DateTime _nextCheck;
-
-		/// <summary>
-		/// How many checks ago were the trails requested
-		/// </summary>
-		private int _trailsAge = 1;
 
 		/// <summary>
 		/// Stop the loop when it's suitable to do so
@@ -58,11 +44,12 @@ namespace PlaneAlerter.Services {
 		/// <summary>
 		/// Constructor
 		/// </summary>
-		public CheckerService(ISettingsManagerService settingsManagerService, IConditionManagerService conditionManagerService,
+		public CheckerService(ISettingsManagerService settingsManagerService, IConditionManagerService conditionManagerService, IVrsService vrsService,
 			IEmailService emailService, ITwitterService twitterService, IUrlBuilderService urlBuilderService, IStatsService statsService)
 		{
 			_settingsManagerService = settingsManagerService;
 			_conditionManagerService = conditionManagerService;
+			_vrsService = vrsService;
 			_emailService = emailService;
 			_twitterService = twitterService;
 			_urlBuilderService = urlBuilderService;
@@ -90,11 +77,11 @@ namespace PlaneAlerter.Services {
 				Core.Ui.UpdateStatusLabel("Downloading Aircraft Info...");
 				receiverName = "";
 				//Get latest aircraft information
-				GetAircraft(false, true);
-				if (_settingsManagerService.Settings.FilterDistance && !_settingsManagerService.Settings.IgnoreModeS) GetAircraft(true, false);
+				_vrsService.GetAircraft(false, true);
+				if (_settingsManagerService.Settings.FilterDistance && !_settingsManagerService.Settings.IgnoreModeS) _vrsService.GetAircraft(true, false);
 
 				//Check if there are aircraft to check
-				if (Core.AircraftList.Count != 0) {
+				if (_vrsService.AircraftList.Count != 0) {
 					//Aircraft number to be shown on UI
 					var aircraftCount = 1;
 					//Current condition being checked
@@ -103,9 +90,10 @@ namespace PlaneAlerter.Services {
 					var ignoreFollowing = false;
 					//Updated trails are available, this check contains the first match in a while
 					var updatedTrailsAvailable = false;
-					foreach (var aircraft in Core.AircraftList.ToList()) {
+					foreach (var aircraft in _vrsService.AircraftList.ToList()) {
 						//Update UI with aircraft being checked
-						Core.Ui.UpdateStatusLabel("Checking conditions for aircraft " + aircraftCount + " of " + Core.AircraftList.Count());
+						Core.Ui.UpdateStatusLabel("Checking conditions for aircraft " + aircraftCount + " of " +
+						                          _vrsService.AircraftList.Count());
 						aircraftCount++;
 
 						if (Core.ActiveMatches.ContainsKey(aircraft.Icao) && Core.ActiveMatches[aircraft.Icao].IgnoreFollowing) continue;
@@ -168,7 +156,8 @@ namespace PlaneAlerter.Services {
 							//Check if condition still matches
 							if (triggersMatching == condition.Triggers.Count) {
 								//Get receiver name
-								if (Core.Receivers.ContainsKey(aircraft.GetProperty("Rcvr"))) receiverName = Core.Receivers[aircraft.GetProperty("Rcvr")];
+								if (_vrsService.Receivers.ContainsKey(aircraft.GetProperty("Rcvr")))
+									receiverName = _vrsService.Receivers[aircraft.GetProperty("Rcvr")];
 
 								//If active matches contains aircraft, add condition to the match
 								if (Core.ActiveMatches.ContainsKey(aircraft.Icao)) {
@@ -186,14 +175,14 @@ namespace PlaneAlerter.Services {
 
 								//Get trails if they haven't been requested due to no matches
 								if (Core.ActiveMatches.Count == 1 && _settingsManagerService.Settings.TrailsUpdateFrequency != 0) {
-									GetAircraft(false, true, true);
-									if (_settingsManagerService.Settings.FilterDistance && !_settingsManagerService.Settings.IgnoreModeS) GetAircraft(true, false, true);
+									_vrsService.GetAircraft(false, true, true);
+									if (_settingsManagerService.Settings.FilterDistance && !_settingsManagerService.Settings.IgnoreModeS) _vrsService.GetAircraft(true, false, true);
 									updatedTrailsAvailable = true;
 								}
 
 								//If updated trails are available, update trail and position
 								if (updatedTrailsAvailable) {
-									foreach (var updatedAircraft in Core.AircraftList.ToList())
+									foreach (var updatedAircraft in _vrsService.AircraftList.ToList())
 									{
 										if (updatedAircraft.Icao != aircraft.Icao)
 											continue;
@@ -247,7 +236,8 @@ namespace PlaneAlerter.Services {
 									condition = c.Condition;
 
 									//Get receiver name
-									if (Core.Receivers.ContainsKey(aircraft.GetProperty("Rcvr"))) receiverName = Core.Receivers[aircraft.GetProperty("Rcvr")];
+									if (_vrsService.Receivers.ContainsKey(aircraft.GetProperty("Rcvr")))
+										receiverName = _vrsService.Receivers[aircraft.GetProperty("Rcvr")];
 
 									//Send Alert
 									SendAlert(condition, aircraft, receiverName, false);
@@ -256,7 +246,7 @@ namespace PlaneAlerter.Services {
 							}
 							//Check if signal has been lost/returned
 							var stillActive = false;
-							foreach (var aircraft in Core.AircraftList)
+							foreach (var aircraft in _vrsService.AircraftList)
 								if (aircraft.Icao == match.Icao)
 									stillActive = true;
 							if (!stillActive && match.SignalLost == false) {
@@ -366,194 +356,53 @@ namespace PlaneAlerter.Services {
 		}
 
 		/// <summary>
-		/// Get latest aircraftlist.json
+		/// Parse a custom format string to replace property names with values
 		/// </summary>
-		private void GetAircraft(bool modeSOnly, bool clearExisting, bool forceRequestTrails = false) {
-			var requestTrails = _settingsManagerService.Settings.TrailsUpdateFrequency==1;
+		/// <returns>Parsed string</returns>
+		public string ParseCustomFormatString(string format, Aircraft aircraft, Condition condition)
+		{
+			var variables = new Dictionary<string, string>
+			{
+				{ "ConditionName", condition.Name },
+				{
+					"RcvrName",
+					_vrsService.Receivers.ContainsKey(aircraft.GetProperty("Rcvr")) ? _vrsService.Receivers[aircraft.GetProperty("Rcvr")] : ""
+				},
+				{ "Date", DateTime.Now.ToString("d") },
+				{ "Time", DateTime.Now.ToString("t") },
+			};
 
-			//Force request trails
-			if (forceRequestTrails) {
-				requestTrails = true;
-			}
-			//No matches so we don't need trails
-			else if (Core.ActiveMatches.Count == 0) {
-				requestTrails = false;
-			}
-			//Threshold enabled
-			else if (_settingsManagerService.Settings.TrailsUpdateFrequency >= 2) {
-				if (_trailsAge >= _settingsManagerService.Settings.TrailsUpdateFrequency) {
-					requestTrails = true;
-					_trailsAge = 0;
+			//Iterate variables
+			foreach (var varKey in variables.Keys)
+			{
+				//Check if content contains keyword
+				if (format.ToLower().Contains(@"[" + varKey.ToLower() + @"]"))
+				{
+					//Replace keyword with value
+					format = Regex.Replace(format, @"\[" + varKey + @"\]", variables[varKey], RegexOptions.IgnoreCase);
 				}
-				_trailsAge++;
 			}
 
-			//Generate aircraftlist.json url
-			var url = _settingsManagerService.Settings.AircraftListUrl;
-			url += _settingsManagerService.Settings.AircraftListUrl.Contains("?") ? "&" : "?";
-			url += "lat=" + _settingsManagerService.Settings.Lat + "&lng=" + _settingsManagerService.Settings.Long;
-			if (_settingsManagerService.Settings.FilterDistance && !modeSOnly) url += "&fDstU=" + _settingsManagerService.Settings.IgnoreDistance.ToString("#.##");
-			if (_settingsManagerService.Settings.FilterAltitude) url += "&fAltU=" + _settingsManagerService.Settings.IgnoreAltitude;
-			if (_settingsManagerService.Settings.FilterReceiver) url += "&feed=" + _settingsManagerService.Settings.FilterReceiverId;
-			if (modeSOnly) url += "&fNoPosQN=1";
-			if (requestTrails) url += "&trFmt=fa&refreshTrails=1";
+			//Iterate properties
+			foreach (var info in Core.VrsPropertyData.Values)
+			{
+				//Check if content contains keyword
+				if (!format.ToLower().Contains(@"[" + info[2].ToLower() + @"]"))
+					continue;
 
-			try {
-				JObject responseJson;
-				try {
-					responseJson = RequestAircraftList(url);
-				}
-				catch (Exception e) {
-					Core.Ui.WriteToConsole("ERROR: " + e.GetType() + " while downloading AircraftList.json: " + e.Message, Color.Red);
-					return;
-				}
+				//Replace keyword with value
+				var value = aircraft.GetProperty(info[2]) ?? "";
 
-				//Check if we actually got aircraft data
-				if (responseJson["acList"] == null) {
-					Core.Ui.WriteToConsole("ERROR: Invalid response received from server", Color.Red);
-					return;
+				//If enum, replace with string value
+				if (EnumUtils.TryGetConvertedValue(info[2], value, out string convertedValue))
+				{
+					value = convertedValue;
 				}
 
-				//Throw error if server time was not parsed
-				if (responseJson["stm"] == null)
-					throw new JsonReaderException();
-
-				//Save old trails if not requesting new ones
-				Dictionary<string, double[]>? oldTrails = null;
-				if (!requestTrails) oldTrails = Core.AircraftList.ToDictionary(x => x.Icao, x => x.Trail);
-
-				//Parse aircraft data
-				if (clearExisting) Core.AircraftList.Clear();
-				foreach (JObject a in responseJson["acList"].ToList()) {
-					//Ignore if no icao is provided
-					if (a["Icao"] == null) continue;
-					//Create new aircraft
-					var aircraft = new Aircraft(a["Icao"].ToString());
-
-					//Parse aircraft trail
-					if (requestTrails) {
-						if (a["Cot"] != null)
-							aircraft.Trail = new double[a["Cot"].Count()];
-						for (var i = 0; i < aircraft.Trail.Length - 1; i++)
-							if (a["Cot"][i].Value<string>() != null)
-								aircraft.Trail[i] = double.Parse(a["Cot"][i].Value<string>(), CultureInfo.InvariantCulture);
-							else
-								aircraft.Trail[i] = 0;
-					}
-					else {
-						if (oldTrails != null && oldTrails.ContainsKey(aircraft.Icao)) aircraft.Trail = oldTrails[aircraft.Icao];
-					}
-
-					//Parse aircraft properties
-					var properties = a.Properties().ToList();
-					for (var i = 0; i < properties.Count; i++)
-						aircraft.AddProperty(properties[i].Name, properties[i].Value.ToString());
-					
-					//Add aircraft to list
-					Core.AircraftList.Add(aircraft);
-				}
-
-				//Get list of receivers
-				Core.Receivers.Clear();
-				foreach (JObject f in responseJson["feeds"])
-					Core.Receivers.Add(f["id"].ToString(), f["name"].ToString());
-
-				//Try to clean up json parsing
-				responseJson.RemoveAll();
-				GC.Collect(2, GCCollectionMode.Forced);
-			}
-			catch (UriFormatException) {
-				Core.Ui.WriteToConsole("ERROR: AircraftList.json url invalid (" + _settingsManagerService.Settings.AircraftListUrl + ")", Color.Red);
-				return;
-			}
-			catch (InvalidDataException) {
-				Core.Ui.WriteToConsole("ERROR: Data returned from " + _settingsManagerService.Settings.AircraftListUrl + " was not gzip compressed", Color.Red);
-				return;
-			}
-			catch (WebException e) {
-				Core.Ui.WriteToConsole("ERROR: Error while connecting to AircraftList.json (" + e.Message + ")", Color.Red);
-				return;
-			}
-			catch (JsonReaderException e) {
-				Core.Ui.WriteToConsole("ERROR: Error parsing JSON response (" + e.Message + ")", Color.Red);
-				return;
-			}
-		}
-
-		private JObject? RequestAircraftList(string url) {
-			//Create request
-			_request = (HttpWebRequest)WebRequest.Create(url);
-			_request.Method = "GET";
-			_request.AutomaticDecompression = DecompressionMethods.Deflate | DecompressionMethods.GZip;
-			_request.Timeout = _settingsManagerService.Settings.Timeout * 1000;
-			//Add credentials if they are provided
-			if (_settingsManagerService.Settings.VRSAuthenticate) {
-				var encoded = Convert.ToBase64String(Encoding.GetEncoding("ISO-8859-1").GetBytes(
-					_settingsManagerService.Settings.VRSUser + ":" + _settingsManagerService.Settings.VRSPassword));
-				_request.Headers.Add("Authorization", "Basic " + encoded);
-			}
-			//Send request and parse json response
-			using var response = (HttpWebResponse)_request.GetResponse();
-			using var responseStream = response.GetResponseStream();
-			using var reader = new StreamReader(responseStream);
-			using var jsonReader = new JsonTextReader(reader);
-
-			return JsonSerializer.Create().Deserialize<JObject>(jsonReader);
-		}
-
-		public Dictionary<string, string>? GetReceivers() {
-			//Generate aircraftlist.json url
-			var url = _settingsManagerService.Settings.AircraftListUrl;
-			url += _settingsManagerService.Settings.AircraftListUrl.Contains("?") ? "&" : "?";
-			url += "fUtQ=abc";
-
-			try {
-				JObject responseJson;
-				try {
-					responseJson = RequestAircraftList(url);
-				}
-				catch (Exception e) {
-					Core.Ui.WriteToConsole("ERROR: " + e.GetType() + " while downloading AircraftList.json: " + e.Message, Color.Red);
-					return null;
-				}
-
-				//Check if we actually got aircraft data
-				if (responseJson["acList"] == null) {
-					Core.Ui.WriteToConsole("ERROR: Invalid response recieved from server", Color.Red);
-					return null;
-				}
-				//Throw error if server time was not parsed
-				if (responseJson["stm"] == null)
-					throw new JsonReaderException();
-
-				//Get list of receivers
-				Core.Receivers.Clear();
-				foreach (JObject f in responseJson["feeds"])
-					Core.Receivers.Add(f["id"].ToString(), f["name"].ToString());
-
-				//Try to clean up json parsing
-				responseJson.RemoveAll();
-
-				GC.Collect(2, GCCollectionMode.Forced);
-			}
-			catch (UriFormatException) {
-				Core.Ui.WriteToConsole("ERROR: AircraftList.json url invalid (" + _settingsManagerService.Settings.AircraftListUrl + ")", Color.Red);
-				return null;
-			}
-			catch (InvalidDataException) {
-				Core.Ui.WriteToConsole("ERROR: Data returned from " + _settingsManagerService.Settings.AircraftListUrl + " was not gzip compressed", Color.Red);
-				return null;
-			}
-			catch (WebException e) {
-				Core.Ui.WriteToConsole("ERROR: Error while connecting to AircraftList.json (" + e.Message + ")", Color.Red);
-				return null;
-			}
-			catch (JsonReaderException e) {
-				Core.Ui.WriteToConsole("ERROR: Error parsing JSON response (" + e.Message + ")", Color.Red);
-				return null;
+				format = Regex.Replace(format, @"\[" + info[2] + @"\]", value, RegexOptions.IgnoreCase);
 			}
 
-			return Core.Receivers;
+			return format;
 		}
 	}
 }
